requirements.txt
playwright==1.58.0
python-telegram-bot==22.7
anthropic==0.94.0
apscheduler==3.11.2
aiosqlite==0.22.1
pydantic-settings==2.13.1
python-dotenv==1.2.2
